seminario: add artifacts related to the seminar submission
A seminar on a paper of interest (to me and the research group) is to be
presented as part of the course activities. Before the presentation, we
must submit the title, summary, and justification of the choice of the
paper.

In this sense, add the dir `seminario` to house the required artifacts.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>